fix: avoid component calling server code
