+++ config/index.ts
@@ -1,2 +1,4 @@
 
 export const DEFAULT_ATTACHED_MESSAGES_COUNT = 10
+
+export * from './models'
